housing_society_management/requirements.txt
Django==2.2.2
django-ajax-selects==1.7.1
django-bootstrap4==0.0.8
django-braces==1.13.0
django-ckeditor==5.7.1
django-crispy-forms==1.8.0
django-datatables-view==1.19.0
django-extensions==2.2.5
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.10.3
djangorestframework-jsonp==1.0.2
docopt==0.6.2
isort==4.3.21
num2words==0.5.10
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
